Tests/LibWeb/Text/expected/HTML/custom-element-parser-reentrant-document-write.txt

connected
written parent: body
host child count: 3
body children: reentrant-writer,span
